refactor: `MemoCard` 컴포넌트 분리
